helmet_detection/requirements.txt
ultralytics>=8.0.0
streamlit>=1.25.0
opencv-python>=4.8.0
pillow>=9.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0 